Add game page (still very buggy).
