+++ src/main.rs
@@ -1,14 +1,14 @@
 mod config;
 mod game;
 mod player;
+mod gui;
 use iced::{window::Settings, Sandbox};
 use std::io;
-mod gui;
 
 fn main() -> Result<(), io::Error> {
     let mut settings = iced::Settings::default();
     settings.window = Settings {
-        size: (640, 480),
+        size: (720, 480),
         ..Settings::default()
     };
     gui::Gui::run(settings);
